Complete iteration2
