+++ src/components/UrlContainer/UrlContainer.js
@@ -1,28 +1,47 @@
-import React from 'react';
+import React, { Component } from 'react';
 import './UrlContainer.css';
 import { connect } from 'react-redux';
 import { setUrls } from '../../actions';
+import { getUrls } from '../../apiCalls';
 
-export const UrlContainer = (props) => {
-  const urlEls = props.urls.map(url => {
+export class UrlContainer extends Component {
+
+  componentDidMount() {
+    getUrls()
+      .then(data => this.props.setUrls(data.urls))
+      .catch(err => console.error('Error fetching:', err));
+  }
+
+  render() {
+
+    const urlEls = this.props.urls.map(url => {
+      return (
+        <div className="url">
+          <h3>{url.title}</h3>
+          <a href={url.short_url} target="blank">{url.short_url}</a>
+          <p>{url.long_url}</p>
+        </div>
+      )
+    });
+
     return (
-      <div className="url">
-        <h3>{url.title}</h3>
-        <a href={url.short_url} target="blank">{url.short_url}</a>
-        <p>{url.long_url}</p>
-      </div>
+      <section>
+        { urlEls.length ? urlEls : <p>No urls yet! Find some to shorten!</p> }
+      </section>
     )
-  });
 
-  return (
-    <section>
-      { urlEls.length ? urlEls : <p>No urls yet! Find some to shorten!</p> }
-    </section>
-  )
+  }
+
 }
 
 export const mapStateToProps = ({ urls }) => ({
   urls
 });
 
-export default connect(mapStateToProps)(UrlContainer);
+export const mapDispatchToProps = dispatch => {
+  return {
+    setUrls: urls => dispatch(setUrls(urls))
+  }
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(UrlContainer);
